Fix: MaterialManager에서 double→float 축소 변환 경고 수정
JSON ToFloat() 반환값(double)을 float 파라미터에 전달할 때
명시적 캐스트 추가.

+++ KraftonEngine/Source/Engine/Materials/MaterialManager.cpp
@@ -174,16 +174,16 @@ void FMaterialManager::ApplyParameters(UMaterial* Material, json::JSON& JsonData
 		{
 			if (Value.length() == 3)
 			{
-				Material->SetVector3Parameter(ParamName, FVector(Value[0].ToFloat(), Value[1].ToFloat(), Value[2].ToFloat()));
+				Material->SetVector3Parameter(ParamName, FVector((float)Value[0].ToFloat(), (float)Value[1].ToFloat(), (float)Value[2].ToFloat()));
 			}
 			else if (Value.length() == 4)
 			{
-				Material->SetVector4Parameter(ParamName, FVector4(Value[0].ToFloat(), Value[1].ToFloat(), Value[2].ToFloat(), Value[3].ToFloat()));
+				Material->SetVector4Parameter(ParamName, FVector4((float)Value[0].ToFloat(), (float)Value[1].ToFloat(), (float)Value[2].ToFloat(), (float)Value[3].ToFloat()));
 			}
 		}
 		else if (Value.JSONType() == json::JSON::Class::Floating || Value.JSONType() == json::JSON::Class::Integral)
 		{
-			Material->SetScalarParameter(ParamName, Value.ToFloat());
+			Material->SetScalarParameter(ParamName, (float)Value.ToFloat());
 		}
 	}
 }
